+++ rally/cli/commands/plugin.py
@@ -18,6 +18,7 @@ from __future__ import print_function
 from rally.cli import cliutils
 from rally.common.plugin import plugin
 from rally.common import utils
+from rally import exceptions
 from rally import plugins
 
 
@@ -58,6 +59,7 @@ class PluginCommands(object):
                 )
             else:
                 print("Plugin %s not found at any platform" % name)
+            return exceptions.PluginNotFound.error_code
 
         elif len(found) == 1 or exact_match:
             plugin_ = found[0] if len(found) == 1 else exact_match[0]
@@ -79,6 +81,7 @@ class PluginCommands(object):
         else:
             print("Multiple plugins found:")
             self._print_plugins_list(found)
+            return exceptions.MultiplePluginsFound.error_code
 
     @cliutils.args(
         "--name", dest="name", type=str,